Fix font-sizes changing with user preference and remove checkbox trick.
The UI would break if the user set their default font size to more than
16px. This was due to the use of 'rem' units in the CSS rather than
'px' units. All UI text now uses 'px'.

The checkbox trick was previously used for toggling the mobile
navigation menu. This was undesirable because it used what was
essentially a dirty hack to show a new element, which may mess with
screen readers. It has been replaced with a <details> element, although
the border is currently bugged.