Compile html with jinja to avoid copy-pasting nav-header everywhere.
.html.jinja files in the src/ folder are now the authoritative files,
though the plain .html files are the ones that are shown on the website.

We now use jinja to find-and-replace common elements like the navigation
header and also the footer once I get around to it.
It will be a pain to hit compile all of the time, but now at least the
header and footer are guaranteed to be consistent across pages and
easier to modify.

Be wary of using jinja too much; only use it for elements that are truly
consisten across pages.