Add hovering descriptions to links.html.
Unfortunately, this change requires the duplication of the description
in the html for each link. There is simply no other way (that I can think
of) to both keep the click-state of details elements for mobile users and
the hover functionality for desktop users. As far as I know, there is no
way to show elements hidden in the details element without modifying the
html.