links.html: Make details wrap onto newlines + other minor changes.
Made tooltips always clear the details box with top: 100%.
Even on mobile, where hovering is impossible, it works.

Also made the infobox larger on mobile.

The panes containing links are set to 300px by the grid layout, which
unfortunately makes the page layout on mobile asymmetrical. I don't
really see a solution to this, but since it's still legible (and
navigating tabs on mobile is a pain anyways), it's fine.